jdv-app/requirements.txt
Kivy==2.3.0
KivyMD==1.2.0
httpx==0.27.0
python-dateutil==2.9.0
kivy_garden.zbarcam
Pillow
plyer